# api/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Smart home api",
    "description": "This documentation describes APIs found under https://github.com/e154/smart-home",
    "version": "1.0",
    "contact": {
      "name": "Alex Filippov",
      "url": "https://e154.github.io/smart-home/",
      "email": "support@e154.ru"
    }
  },
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/task": {
      "post": {
        "summary": "add new task",
        "operationId": "AutomationService_AddTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTask"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiNewTaskRequest"
            }
          }
        ],
        "tags": [
          "AutomationService"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/task/{id}": {
      "get": {
        "summary": "get task",
        "operationId": "AutomationService_GetTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTask"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "AutomationService"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "summary": "delete task",
        "operationId": "AutomationService_DeleteTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "AutomationService"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "put": {
        "summary": "update task",
        "operationId": "AutomationService_UpdateTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTask"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "enabled": {
                  "type": "boolean"
                },
                "condition": {
                  "$ref": "#/definitions/apiUpdateTaskRequestConditionEnum"
                },
                "triggers": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/apiUpdateTaskRequestTrigger"
                  }
                },
                "conditions": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/apiUpdateTaskRequestCondition"
                  }
                },
                "actions": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/apiUpdateTaskRequestAction"
                  }
                },
                "area": {
                  "$ref": "#/definitions/apiArea"
                }
              }
            }
          }
        ],
        "tags": [
          "AutomationService"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/tasks": {
      "get": {
        "summary": "get task list",
        "operationId": "AutomationService_GetTaskList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetTaskListResult"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "AutomationService"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "apiAction": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "taskId": {
          "type": "string",
          "format": "int64"
        },
        "script": {
          "$ref": "#/definitions/apiScript"
        }
      }
    },
    "apiArea": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
...
    "apiGetTaskListResult": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiTask"
          }
        },
        "meta": {
          "$ref": "#/definitions/apiMeta"
        }
      }
    },
...
    "apiNewTaskRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        },
        "condition": {
          "$ref": "#/definitions/apiNewTaskRequestConditionEnum"
        },
        "triggers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiNewTaskRequestTrigger"
          }
        },
        "conditions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiNewTaskRequestCondition"
          }
        },
        "actions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiNewTaskRequestAction"
          }
        },
        "area": {
          "$ref": "#/definitions/apiArea"
        }
      }
    },
...
    "apiNewTaskRequestCondition": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "taskId": {
          "type": "string",
          "format": "int64"
        },
        "scriptId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "apiNewTaskRequestConditionEnum": {
      "type": "string",
      "enum": [
        "OR",
        "AND"
      ],
      "default": "OR"
    },
    "apiNewTaskRequestTrigger": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "taskId": {
          "type": "string",
          "format": "int64"
        },
        "entityId": {
          "type": "string"
        },
        "scriptId": {
          "type": "string",
          "format": "int64"
        },
        "pluginName": {
          "type": "string"
        },
        "payload": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/apiAttribute"
          }
        }
      }
    },
...
    "apiTask": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        },
        "condition": {
          "$ref": "#/definitions/apiTaskCondition"
        },
        "triggers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiTrigger"
          }
        },
        "conditions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiTaskCondition"
          }
        },
        "actions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiAction"
          }
        },
        "area": {
          "$ref": "#/definitions/apiArea"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "apiTaskCondition": {
      "type": "string",
      "enum": [
        "OR",
        "AND"
      ],
      "default": "OR"
    },
    "apiTrigger": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "taskId": {
          "type": "string",
          "format": "int64"
        },
        "entityId": {
          "type": "string"
        },
        "script": {
          "$ref": "#/definitions/apiScript"
        },
        "pluginName": {
          "type": "string"
        },
        "payload": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/apiAttribute"
          }
        }
      }
    },
...
    "apiUpdateTaskRequestAction": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "taskId": {
          "type": "string",
          "format": "int64"
        },
        "scriptId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "apiUpdateTaskRequestCondition": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "taskId": {
          "type": "string",
          "format": "int64"
        },
        "scriptId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "apiUpdateTaskRequestConditionEnum": {
      "type": "string",
      "enum": [
        "OR",
        "AND"
      ],
      "default": "OR"
    },
    "apiUpdateTaskRequestTrigger": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "taskId": {
          "type": "string",
          "format": "int64"
        },
        "entityId": {
          "type": "string"
        },
        "scriptId": {
          "type": "string",
          "format": "int64"
        },
        "pluginName": {
          "type": "string"
        },
        "payload": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/apiAttribute"
          }
        }
      }
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "x-api-key",
      "in": "header",
      "x-api-key": "jwt"
    },
    "BasicAuth": {
      "type": "basic"
    }
  },
  "security": [
    {
      "ApiKeyAuth": [],
      "BasicAuth": []
    }
  ]
}